+++ lib/collaboration_web/controllers/admin_controller.ex
@@ -1,6 +1,0 @@
-defmodule CollaborationWeb.AdminController do
-  use CollaborationWeb, :controller
-
-  plug(CollaborationWeb.Plug.IsAdmin)
-  def users(conn, _params), do: render(conn, "users.html")
-end