+++ lib/collaboration_web/controllers/page_controller.ex
@@ -1,5 +1,0 @@
-defmodule CollaborationWeb.PageController do
-  use CollaborationWeb, :controller
-
-  def index(conn, _), do: redirect(conn, to: topic_path(conn, :index))
-end