+++ lib/collaboration_web/controllers/topic_controller.ex
@@ -1,26 +1,14 @@
 defmodule CollaborationWeb.TopicController do
   use CollaborationWeb, :controller
 
-  import Collaboration.Contributions
   import CollaborationWeb.ViewHelpers, only: [admin?: 1]
 
   def index(conn, _) do
     render conn, "index.html", topics: list_topics(admin?(conn))
-    # topics = list_topics(admin?(conn))
-    # render(assign(conn, :topics, topics), "index.html")
   end
 
-  def show(conn, %{"id" => id} = params) do
-    current_user = Coherence.current_user(conn)
-    topic = get_topic!(id)
-    ideas = list_ideas(topic.id, current_user)
-    current_idea = Map.get(params, "idea", nil)
-    render conn, "show.html",
-      current_idea: Enum.find(ideas, fn(x) -> x.id == current_idea end),
-      current_user: current_user,
-      topic: topic,
-      ideas: ideas,
-      my_ratings: list_my_ratings(Enum.map(ideas, & &1.id), current_user)
+  def show(conn, %{"id" => id}) do
+    redirect conn, to: topic_idea_path(conn, :index, id)
   end
 
   def new(conn, _), do: render(conn, "new.html", changeset: change_topic())