+++ lib/collaboration_web/views/comment_view.ex
@@ -1,20 +1,22 @@
 defmodule CollaborationWeb.CommentView do
   use CollaborationWeb, :view
 
-  def render("comment.json", %{comment: c, current_user: user_id}) do
-    liked_ids = Enum.map(c.likes, fn l -> l.name end)
-    liked = if user_id, do: Enum.member?(liked_ids, user_id), else: false
+  def thumbsUp?(comment, user) do
+    if comment.likes > 0 || (user && (comment.id !== user.id || user.admin)),
+    do: true, else: false
+  end
 
+  def render("comment.json", %{comment: c}) do
+    liked = !!c.liked
+    likes = if liked, do: c.likes + 1, else: c.likes
     %{
-      author: c.user.name,
-      recipient_id: c.recipient_id,
       id: c.id,
+      author: c.author,
+      created: NaiveDateTime.to_iso8601(c.created) <> "Z",
       text: c.text,
-      created: NaiveDateTime.to_iso8601(c.inserted_at) <> "Z",
-      fake_likes: c.fake_likes,
-      likes: c.fake_likes + Enum.count(liked_ids),
-      liked: liked,
-      public: c.public
+      liked: !!c.liked,
+      likes: likes,
+      user_id: c.user_id
     }
   end
 end