+++ lib/collaboration_web/views/view_helpers.ex
@@ -4,6 +4,7 @@ defmodule CollaborationWeb.ViewHelpers do
   """
   use Phoenix.HTML
   import Phoenix.Controller, only: [current_path: 2]
+  import Coherence, only: [current_user: 1]
 
   @type conn :: Plug.Conn.t()
 
@@ -20,8 +21,9 @@ defmodule CollaborationWeb.ViewHelpers do
 
   def icon(class), do: content_tag(:i, "", class: class)
 
-  def user?(conn), do: !!Coherence.current_user(conn)
-  def admin?(conn), do: user?(conn) && Coherence.current_user(conn).admin
+  def user?(conn), do: !!current_user(conn)
+  def admin?(conn), do: user?(conn) && current_user(conn).admin
+  def user_id(conn), do: user?(conn) && current_user(conn).id
 
   def nav_text(text, icon \\ false) do
     text = if icon, do: [icon(icon), text], else: text